Update README.md
Nested Loops

A loop can be nested inside of another loop. Write a program that uses nested loops to collect data and calculate the average rainfall over a period of years.

The program should first ask for the number of years. The outer loop will iterate once for each year. The inner loop will iterate four times, once for each quarter. Each iteration of the inner loop will ask the user for the inches of rainfall for that quarter.

After all iterations, the program should display the number of months, the total inches of rainfall, and the average rainfall per month for the entire period.

Input Validation: Do not accept a number less than 1 for the number of years. Do not accept negative numbers for the quarterly rainfall.